add policy

+++ resources/views/todos/index.blade.php
@@ -44,9 +44,15 @@
                                     <td> {{ $todo->created_at->diffForHumans() }}</td>
                                     <td> {{ $todo->user->name }}</td>
                                     <td>
-                                        <a class="btn btn-primary" href="/todos/{{ $todo->id}}">Show</a>
-                                        <a class="btn btn-success" href="/todos/{{ $todo->id}}/edit">Edit</a>
-                                        <a onclick="return confirm('Are you sure')" class="btn btn-danger" href="/todos/{{ $todo->id}}/delete">Delete</a>
+                                        @can('lihat',$todo)
+                                            <a class="btn btn-primary" href="/todos/{{ $todo->id}}">Show</a>
+                                        @endcan
+                                        @can('edit', $todo)
+                                            <a class="btn btn-success" href="/todos/{{ $todo->id}}/edit">Edit</a>
+                                        @endcan
+                                        @can('delete', $todo)
+                                            <a onclick="return confirm('Are you sure')" class="btn btn-danger" href="/todos/{{ $todo->id}}/delete">Delete</a>
+                                        @endcan
                                     </td>
                                 </tr>
                             @endforeach
